combined parsers in one file
